src/requirements.txt
Django==1.3
Pygments==1.4
distribute==0.6.10
django-debug-toolbar==0.8.5
django-mptt==0.4.2
gunicorn==0.14.2
psycopg2==2.4
python-memcached==1.47
wsgiref==0.1.2
